+++ NEWS.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.14] – 2024-01-31
+
+### Added
+
+* `rand` on `UnitaryMatrices`
+* `rand` on arbitrary `GroupManifold`s and manifolds with `IsGroupManifold` trait
+  generating points and elements from the Lie algebra, respectively
+
 ## [0.9.13] – 2024-01-24
 
 ### Added